+++ CMakeLists.txt
@@ -28,7 +28,6 @@ if (APPLE)
 endif ()
 
 if (UNIX AND NOT APPLE)
-  add_subdirectory(src/main/backends/linux/niri/native)
   add_subdirectory(src/main/backends/linux/wlroots/native)
   add_subdirectory(src/main/backends/linux/x11/native)
 endif ()
