+++ locales/en/translation.json
@@ -401,7 +401,7 @@
       "shortcut-info": "Kando cannot directly bind global shortcuts on KDE/Wayland. Instead, specify a shortcut ID here and then search for it in the KWin section of the global shortcuts in your KDE's system settings. There you can assign the shortcut which will open this menu."
     },
     "niri": {
-      "shortcut-info": "Kando can directly bind global shortcuts on Niri. However, Niri doesn't implement triggering them. It might in the future."
+      "shortcut-info": "Niri does not yet support global shortcuts. But you can open this menu using 'kando --menu \"Menu Name\"' from the command line or from your niri config."
     }
   }
 }
