+++ src/main/backends/linux/wlroots/backend.ts
@@ -61,4 +61,12 @@ export abstract class WLRBackend extends Backend {
       native.simulateKey(keyCodes[i], keys[i].down);
     }
   }
+  
+  /**
+   * This gets the pointer's position and work area size without relying on IPCs by
+   * spawning an wlr_layer_shell overlay surface
+   */
+  protected getPointerPositionAndWorkAreaSize() {
+    return native.getPointerPositionAndWorkAreaSize();
+  }
 }
